+++ Questionable/Windows/QuestSelectionWindow.cs
@@ -226,11 +226,9 @@ internal sealed class QuestSelectionWindow : LWindow
                 else if (ImGui.IsItemClicked(ImGuiMouseButton.Right))
                     CopyToClipboard(quest, false);
                 ImGui.SameLine();
-#if DEBUG
                 if (ImGuiComponentsLocal.IconButton(FontAwesomeIcon.Edit))
                     (bool success, string filename) = QuestRegistry.OpenEditor(quest);
                 ImGui.SameLine();
-#endif
 
                 if (knownQuest != null &&
                     knownQuest.FindSequence(0)?.LastStep()?.InteractionType is EInteractionType.AcceptQuest &&